requirements.txt
sumy==0.7.0
nltk==3.3
PyPDF2==1.26.0
beautifulsoup4==4.6.3
bs4==0.0.1
lxml==4.2.5
urllib3==1.24
wrapt==1.10.11
Django==2.1.4
Jinja2==2.10
MarkupSafe==1.1.0
PyPDF2==1.26.0
breadability==0.1.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.2
cytoolz==0.9.0.1
defusedxml==0.5.0
dill==0.2.8.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djoser==1.3.3
docopt==0.6.2
idna==2.8
itypes==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nameparser==1.0.2
numpy==1.15.4
oauthlib==2.1.0
openapi-codec==1.3.2
pandas==0.23.4
pip==18.1
plac==0.9.6
preshed==2.0.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
regex==2018.1.10
requests==2.21.0
requests-oauthlib==1.0.0
scikit-learn==0.20.2
scipy==1.2.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.29.0
ujson==1.35
uritemplate==3.0.0
certifi==2018.11.29
psycopg2==2.7.6.1
psycopg2-binary==2.7.7